move investigator tests to their own file; add two investigator tests to check latest changes to the endpoint
